fix(#166): address review issues in Phase 1 entity classes
- Replace LocaleString with Map<String, String> in ServiceJob and
  PackingTargetContainer to match the flat JSON shape the API sends
- Delete LocaleString (unused after the fix above)
- Add missing returnedLineItems field to ItemReturn
- Fix Parcel.java to use import rather than fully-qualified java.time.Instant
- Normalise all version fields to Integer (ExpiryEntity.version,
  ExpiryEntity.processVersion, InboundDelivery.version, PurchaseOrder.version)

+++ sdk-core/src/main/java/de/joesst/dev/fulfillmenttools/expiryentities/ExpiryEntity.java
@@ -25,14 +25,14 @@ import java.time.Instant;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public record ExpiryEntity(
         ExpiryEntityId id,
-        Long version,
+        Integer version,
         Instant created,
         Instant lastModified,
         String processRef,
         Instant provisioningTime,
         Instant expiryTime,
         ExpiryEntityStatus status,
-        Long processVersion
+        Integer processVersion
 ) {
 
     public static Builder builder() {
@@ -43,21 +43,21 @@ public record ExpiryEntity(
         private Builder() {}
 
         private ExpiryEntityId id;
-        private Long version;
+        private Integer version;
         private Instant created;
         private Instant lastModified;
         private String processRef;
         private Instant provisioningTime;
         private Instant expiryTime;
         private ExpiryEntityStatus status;
-        private Long processVersion;
+        private Integer processVersion;
 
         public Builder id(ExpiryEntityId id) {
             this.id = id;
             return this;
         }
 
-        public Builder version(Long version) {
+        public Builder version(Integer version) {
             this.version = version;
             return this;
         }
@@ -92,7 +92,7 @@ public record ExpiryEntity(
             return this;
         }
 
-        public Builder processVersion(Long processVersion) {
+        public Builder processVersion(Integer processVersion) {
             this.processVersion = processVersion;
             return this;
         }
